Revert "Refactor to allow RL models"
